moved files to utils directory
