+++ supersuit/aec_wrappers.py
@@ -1,8 +1,8 @@
 from .base_aec_wrapper import BaseWrapper, PettingzooWrap
 from gym.spaces import Box, Space, Discrete
-from . import basic_transforms
+from .utils import basic_transforms
 from .utils.frame_stack import stack_obs_space, stack_init, stack_obs
-from .action_transforms import homogenize_ops
+from .utils.action_transforms import homogenize_ops
 from .utils import agent_indicator as agent_ider
 from .utils.frame_skip import check_transform_frameskip
 from .utils.obs_delay import Delayer
