+++ supersuit/utils/basic_transforms/color_reduction.py
