+++ test/test_utils/test_agent_indicator.py
