+++ test/test_utils/test_basic_transforms/test_dtype.py
@@ -1,4 +1,4 @@
-from supersuit.basic_transforms.flatten import check_param, change_observation
+from supersuit.utils.basic_transforms.dtype import check_param, change_observation
 from gym.spaces import Box
 import numpy as np
 import pytest
@@ -8,11 +8,12 @@ test_obs = np.zeros([4, 4, 3], dtype=np.float64) + np.arange(3)
 
 
 def test_param_check():
-    check_param(test_obs_space, True)
-    with pytest.raises(AssertionError):
+    check_param(test_obs_space, np.uint8)
+    check_param(test_obs_space, np.dtype("uint8"))
+    with pytest.raises(TypeError):
         check_param(test_obs_space, 6)
 
 
 def test_change_observation():
-    new_obs = change_observation(test_obs, test_obs_space, True)
-    assert new_obs.shape == (4 * 4 * 3,)
+    new_obs = change_observation(test_obs, test_obs_space, np.float32)
+    assert new_obs.dtype == np.float32
